requirements.txt
Flask-WTF
Flask-Bootstrap4